Linux port: fix direct-express/express-gpu build on Linux - configure: use -lglfw -lGL instead of Windows glfw3/opengl32 - hw: add egl_window_glx.c (Linux EGL/GLX window backend) - guard Windows-only code (WIN32/CreateEvent/CloseHandle) in egl_surface,   edu.c, express_gpu_render - fix uint64_t in direct_express_distribute - include: add GLFW_INCLUDE_NONE guards before glad headers - add run-atv.sh launcher script (virgl + SDL display)

+++ hw/direct-express/direct_express_distribute.c
@@ -457,7 +457,7 @@ static Direct_Express_Call *pack_call_from_queue(VirtQueue *vq)
     return NULL;
 }
 
-Thread_Context *thread_context_create(unsigned long long thread_id, unsigned long long type_id, unsigned long long len, Express_Device_Info *info)
+Thread_Context *thread_context_create(uint64_t thread_id, uint64_t type_id, uint64_t len, Express_Device_Info *info)
 {
 
     Thread_Context *context = g_malloc(len);
